+++ homelette/alignment.py
@@ -694,16 +694,16 @@ class Alignment():
         66.67
 
         Normalization happens for the length of sequence 1, so the order of
-        sequences matters
+        sequences matters.
 
         >>> aln = hm.Alignment(None)
         >>> aln.sequences = {
         ...     'seq1': hm.alignment.Sequence('seq1', 'AAAACCCCDDDD'),
-        ...     'seq3': hm.alignment.Sequence('seq3', 'AAAA----DDDD')
+        ...     'seq2': hm.alignment.Sequence('seq3', 'AAAA----DDDD')
         ...     }
         >>> aln.calc_identity('seq1', 'seq2')
         66.67
-        >>> aln.calc_identity('seq3', 'seq1')
+        >>> aln.calc_identity('seq2', 'seq1')
         100.0
         '''
 
@@ -762,10 +762,11 @@ class Alignment():
         # iterate over all pairs of sequences
         for sequence_name_1, sequence_name_2 in itertools.product(
                 self.sequences.keys(), repeat=2):
-            output['sequence_1'].append(sequence_name_1)
-            output['sequence_2'].append(sequence_name_2)
-            output['identity'].append(self.calc_identity(
-                sequence_name_1, sequence_name_2))
+            if not sequence_name_1 == sequence_name_2:
+                output['sequence_1'].append(sequence_name_1)
+                output['sequence_2'].append(sequence_name_2)
+                output['identity'].append(self.calc_identity(
+                    sequence_name_1, sequence_name_2))
         return pd.DataFrame(output)
 
     def calc_identity_target(
@@ -800,10 +801,11 @@ class Alignment():
         '''
         output = {'sequence_1': [], 'sequence_2': [], 'identity': []}
         for sequence_name_2 in self.sequences.keys():
-            output['sequence_1'].append(sequence_name)
-            output['sequence_2'].append(sequence_name_2)
-            output['identity'].append(self.calc_identity(
-                sequence_name, sequence_name_2))
+            if not sequence_name == sequence_name_2:
+                output['sequence_1'].append(sequence_name)
+                output['sequence_2'].append(sequence_name_2)
+                output['identity'].append(self.calc_identity(
+                    sequence_name, sequence_name_2))
         return pd.DataFrame(output)
 
     def calc_coverage(self, sequence_name_1: str,