+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Fixed mistake in docstring for `alignment.Alignment.calc_identity()`.
 
+### Changed
+- `alignment.Alignment.calc_identity_target()` and
+`alignment.Alignment.calc_pairwise_identity_all()` do not calculate
+identities between sequences of the same identifier anymore.
+
 ## [1.2] - 2021-09-15
 ### Added
 - Started using a CHANGELOG.