+++ CHANGELOG.md
@@ -12,6 +12,15 @@ potential templates and generating alignments using different methods:
  - `alignment.Alignment.calc_coverage`
  - `alignment.Alignment.calc_coverage_target`
  - `alignment.Alignment.calc_pairwise_coverage_all`
+- Added .gitignore file.
+
+### Fixed
+- Fixed mistake in docstring for `alignment.Alignment.calc_identity()`.
+
+### Changed
+- `alignment.Alignment.calc_identity_target()` and
+`alignment.Alignment.calc_pairwise_identity_all()` do not calculate
+identities between sequences of the same identifier anymore.
 
 ## [1.2] - 2021-09-15
 ### Added